+++ libsolidity/formal/Z3Interface.cpp
@@ -54,18 +54,20 @@ void Z3Interface::declareVariable(string const& _name, Sort const& _sort)
 {
 	if (_sort.kind == Kind::Function)
 		declareFunction(_name, _sort);
-	else if (!m_constants.count(_name))
-		m_constants.insert({_name, m_context.constant(_name.c_str(), z3Sort(_sort))});
+	else if (m_constants.count(_name))
+		m_constants.at(_name) = m_context.constant(_name.c_str(), z3Sort(_sort));
+	else
+		m_constants.emplace(_name, m_context.constant(_name.c_str(), z3Sort(_sort)));
 }
 
 void Z3Interface::declareFunction(string const& _name, Sort const& _sort)
 {
 	solAssert(_sort.kind == smt::Kind::Function, "");
-	if (!m_functions.count(_name))
-	{
-		FunctionSort fSort = dynamic_cast<FunctionSort const&>(_sort);
-		m_functions.insert({_name, m_context.function(_name.c_str(), z3Sort(fSort.domain), z3Sort(*fSort.codomain))});
-	}
+	FunctionSort fSort = dynamic_cast<FunctionSort const&>(_sort);
+	if (m_functions.count(_name))
+		m_functions.at(_name) = m_context.function(_name.c_str(), z3Sort(fSort.domain), z3Sort(*fSort.codomain));
+	else
+		m_functions.emplace(_name, m_context.function(_name.c_str(), z3Sort(fSort.domain), z3Sort(*fSort.codomain)));
 }
 
 void Z3Interface::addAssertion(Expression const& _expr)