+++ scripts/tests.sh
@@ -36,11 +36,6 @@ source "${REPO_ROOT}/scripts/common.sh"
 WORKDIR=`mktemp -d`
 CMDLINE_PID=
 
-if [[ "$OSTYPE" == "darwin"* ]]
-then
-    SMT_FLAGS="--no-smt"
-fi
-
 cleanup() {
     # ensure failing commands don't cause termination during cleanup (especially within safe_kill)
     set +e