+++ Changelog.md
@@ -40,6 +40,8 @@ Compiler Features:
  * Code Generator: Use SELFBALANCE for ``address(this).balance`` if using Istanbul EVM
  * SMTChecker: Add break/continue support to the CHC engine.
  * SMTChecker: Support assignments to multi-dimensional arrays and mappings.
+ * SMTChecker: Support inheritance and function overriding.
+ * EWasm: Experimental EWasm binary output.
 
 
 Bugfixes: